test: Add unit tests for cart Sentry tags and log terminal error handling
